+++ TunnelDemo.cpp
@@ -139,13 +139,13 @@ void TunnelDemo::update(double elapsedTime)
 {
 	static float bla = 0;
 
-	bla+=0.01f * elapsedTime / 60.0f;
+	bla+=(float)(0.01f * elapsedTime / 60.0f);
 	if(wobble)
 		rot = sin(bla) * 0.75f;
 	
 	if(moving)
 	{
-		dist += 0.02f * elapsedTime / 60.0f;
+		dist += (float)(0.02f * elapsedTime / 60.0f);
 		while(dist >= 1.5f)
 			dist-=1.5f;
 	}
